fix: graceful worker exits for same worker jobs (#4371)
* all

* fix compile
