+++ frontend/src/lib/components/graph/renderers/edges/BaseEdge.svelte
@@ -55,7 +55,7 @@
 	{#if data?.insertable && !$useDataflow && !data?.moving}
 		<div
 			class={twMerge('edgeButtonContainer nodrag nopan top-0', menuOpen ? 'z-50' : '')}
-			style:transform="translate(-50%, 50%) translate({sourceX}px,{sourceY}px)"
+			style:transform="translate(-50%, 50%) translate({sourceX}px,{sourceY + 2}px)"
 		>
 			<InsertModuleButton
 				disableAi={data.disableAi}
@@ -79,7 +79,7 @@
 		{#if data.enableTrigger}
 			<div
 				class="edgeButtonContainer nodrag nopan"
-				style:transform="translate(100%, 50%) translate({sourceX}px,{sourceY}px)"
+				style:transform="translate(100%, 50%) translate({sourceX}px,{sourceY + 2}px)"
 			>
 				<InsertTriggerButton
 					disableAi={data.disableAi}
